requirements-dev.txt
pytest==8.2.1
pyyaml==6.0.2
pydantic==2.8.2
ruff==0.6.3
black==24.8.0
isort==5.13.2
mypy==1.11.2
types-PyYAML==6.0.12.20240808
